Moves ts schema files to own dir
